rdma: Use uint16_t for rail indexes
Rail ids and number of rails are always greater than zero.  For
some reason, we used int almost everywhere.  Fix that to always
use uint16_t (assuming we'll never have 256 NICs / GPU).

+++ include/nccl_ofi_scheduler.h
@@ -19,7 +19,7 @@
  */
 typedef struct nccl_net_ofi_xfer_info {
 	/* Id of the rail */
-	int rail_id;
+	uint16_t rail_id;
 	/* Offset of the stripe into the message */
 	size_t offset;
 	/* Size of the stripe in bytes */
